move `ReadLink` to submodule
