+++ src/rust/engine/src/nodes/mod.rs
@@ -44,10 +44,12 @@ use workunit_store::{in_workunit, Level, RunningWorkunit};
 
 // Sub-modules for the differnt node kinds.
 mod execute_process;
+mod read_link;
 mod root;
 
 // Re-export symbols for each kind of node.
 pub use self::execute_process::{ExecuteProcess, ProcessResult};
+pub use self::read_link::{LinkDest, ReadLink};
 pub use self::root::Root;
 
 tokio::task_local! {
@@ -243,38 +245,6 @@ pub fn lift_directory_digest(digest: &PyAny) -> Result<DirectoryDigest, String> 
 pub fn lift_file_digest(digest: &PyAny) -> Result<hashing::Digest, String> {
     let py_file_digest: externs::fs::PyFileDigest = digest.extract().map_err(|e| format!("{e}"))?;
     Ok(py_file_digest.0)
-}
-
-///
-/// A Node that represents reading the destination of a symlink (non-recursively).
-///
-#[derive(Clone, Debug, DeepSizeOf, Eq, Hash, PartialEq)]
-pub struct ReadLink(Link);
-
-impl ReadLink {
-    async fn run_node(self, context: Context) -> NodeResult<LinkDest> {
-        let node = self;
-        let link_dest = context
-            .core
-            .vfs
-            .read_link(&node.0)
-            .await
-            .map_err(|e| throw(format!("{e}")))?;
-        Ok(LinkDest(link_dest))
-    }
-}
-
-#[derive(Clone, Debug, DeepSizeOf, Eq, PartialEq)]
-pub struct LinkDest(PathBuf);
-
-impl CompoundNode<NodeKey> for ReadLink {
-    type Item = LinkDest;
-}
-
-impl From<ReadLink> for NodeKey {
-    fn from(n: ReadLink) -> Self {
-        NodeKey::ReadLink(n)
-    }
 }
 
 ///
@@ -1398,17 +1368,6 @@ impl TryFrom<NodeOutput> for store::Snapshot {
     fn try_from(nr: NodeOutput) -> Result<Self, ()> {
         match nr {
             NodeOutput::Snapshot(v) => Ok(v),
-            _ => Err(()),
-        }
-    }
-}
-
-impl TryFrom<NodeOutput> for LinkDest {
-    type Error = ();
-
-    fn try_from(nr: NodeOutput) -> Result<Self, ()> {
-        match nr {
-            NodeOutput::LinkDest(v) => Ok(v),
             _ => Err(()),
         }
     }
